run sync and remove productoverview
