+++ src/views/zesty/Footer.js
@@ -1,0 +1,1 @@
+export default {};
