+++ src/views/zesty/index.js
@@ -66,69 +66,68 @@ import Header from './Header';
 import Footer from './Footer';
 
 export {
-
-    Clipping,
-    Widget,
-    MarketingAnnoucement,
-    Review,
-    MainNavigation,
-    PersonaContent,
-    CustomerBrand,
-    CustomerHighlight,
-    LandingPageVariation,
-    TeamMember,
-    Feature,
-    IntegrationPartner,
-    IntegrationType,
-    PricingTier,
-    Partner,
-    Hero,
-    PlatformSection,
-    NewNavigation,
-    DeveloperResourceCard,
-    ClientCard,
-    AgencyPartnerCard,
-    AboutStat,
-    JobListing,
-    EventHost,
-    NewsArticleAuthor,
-    DemoPageItem,
-    Cta,
-    Benefit,
-    ProofPoint,
-    PlatformOverviewCard,
-    FeaturesTile,
-    TechnologyBenefit,
-    Category,
-    Article,
-    Tag,
-    Author,
-    UseCase,
-    LandingPage,
-    ProductExplainer,
-    Event,
-    Homepage,
-    SimplePage,
-    Mindshare,
-    MindshareAuthor,
-    About,
-    SalesPeople,
-    Product,
-    Integration,
-    Pricing,
-    Ourpartner,
-    HeroHompage,
-    WhyZesty,
-    PlatformOverview,
-    TechnologyOverview,
-    CaseStudy,
-    AgencyPartner,
-    Client,
-    Contact,
-    Career,
-    DeveloperResource,
-    EventsAndWebinar,
-    Demo,
-    Footer,
-    Header
-}
+  Clipping,
+  Widget,
+  MarketingAnnoucement,
+  Review,
+  MainNavigation,
+  PersonaContent,
+  CustomerBrand,
+  CustomerHighlight,
+  LandingPageVariation,
+  TeamMember,
+  Feature,
+  IntegrationPartner,
+  IntegrationType,
+  PricingTier,
+  Partner,
+  Hero,
+  PlatformSection,
+  NewNavigation,
+  DeveloperResourceCard,
+  ClientCard,
+  AgencyPartnerCard,
+  AboutStat,
+  JobListing,
+  EventHost,
+  NewsArticleAuthor,
+  DemoPageItem,
+  Cta,
+  Benefit,
+  ProofPoint,
+  PlatformOverviewCard,
+  FeaturesTile,
+  TechnologyBenefit,
+  Category,
+  Article,
+  Tag,
+  Author,
+  UseCase,
+  LandingPage,
+  ProductExplainer,
+  Event,
+  Homepage,
+  SimplePage,
+  Mindshare,
+  MindshareAuthor,
+  About,
+  SalesPeople,
+  Product,
+  Integration,
+  Pricing,
+  Ourpartner,
+  HeroHompage,
+  WhyZesty,
+  PlatformOverview,
+  TechnologyOverview,
+  CaseStudy,
+  AgencyPartner,
+  Client,
+  Contact,
+  Career,
+  DeveloperResource,
+  EventsAndWebinar,
+  Demo,
+  Footer,
+  Header,
+};
